refactor: rename pkg names
