+++ command/help.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	_ "embed"
 
-	"github.com/nurali-techie/rc/cli"
 	"github.com/nurali-techie/rc/io"
 )
 
@@ -15,7 +14,7 @@ type helpCommand struct {
 	output io.Output
 }
 
-func NewHelpCommand(output io.Output) cli.Command {
+func NewHelpCommand(output io.Output) Command {
 	helpCmd := new(helpCommand)
 	helpCmd.output = output
 	return helpCmd
