+++ domain/command.go
@@ -1,6 +1,9 @@
 package domain
 
+// Command entity
 type Commmand struct {
 	Key  string
 	Text string
 }
+
+
